requirements.txt
beautifulsoup4==4.12.2
pandas==2.0.3
python-dotenv==1.0.0
Requests==2.31.0
